Crate pallet_price_feed

Price Feed Pallet

This pallet is an abstraction over the chainlink-feed-pallet which provides oracle data from the chainlink network. This requires some more configurations and provides prices for assets used in the index. For the purpose of the PINT Index all prices will be using a single denominating asset which will be one base currency, which maybe USD, so that the net asset value of the index can be calculated. It is therefore assumed that all the registered chainlink feeds are price pairs with a consisting asset price (e.g. USD as in USD/DOT). NOTE: Most chainlink price feeds use USD as the quote currency to easily calculate how much USD is needed to purchase one units of the base currency, or the value of a certain amount of assets by multiplying it with the units the assets. Therefore ths price feed pallet sticks to the same convention, so that the NAV of the index is the sum of all the assets multiplied with their price in form of (Asset/USD) divided by the total supply of index tokens which essentially is the currency price pair of (PINT/USD).

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.