Module pallet_price_feed::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

AssetPricePair

Defines an asset pair identifier

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

TimestampedValue

Represents an answer of a feed at a certain point of time

_GeneratedPrefixForStorageAssetFeeds

_GeneratedPrefixForStorageLatestAnswerTimestamp

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Provides access to all the price feeds This is used to determine the equivalent amount of PINT for assets

PriceFeed

An interface to access price data

PriceFeedBenchmarks

WeightInfo

Trait for the asset-index pallet extrinsic weights.

Type Definitions

AssetFeeds

Store a mapping (AssetId) -> FeedId for all active assets

FeedIdFor

FeedValueFor

LatestAnswerTimestamp

Stores the timestamp of the latest answer of each feed (feed) -> Timestamp

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

MomentOf

Price

The type to represent asset prices

TimestampedFeedValue