Trait pallet_price_feed::pallet::Config

pub trait Config: Config + Config + MaybeAssetIdConvert<u8, Self::AssetId> {
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type SelfAssetId: Get<Self::AssetId>;
    type AssetId: Parameter + Member + MaybeSerializeDeserialize;
    type Time: Time;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
}

Provides access to all the price feeds This is used to determine the equivalent amount of PINT for assets

The internal chainlink oracle type FeedOracle gives access to the asset’s price feeds.

NOTE: this assumes all the feeds provide data in the same base currency. When querying the price of an asset (quote/asset) from the oracle, its price is given by means of the asset pair (base / quote). (e.g. DOT/PINT)

Associated Types

type AdminOrigin: EnsureOrigin<Self::Origin>

The origin that is allowed to insert asset -> feed mappings

type SelfAssetId: Get<Self::AssetId>

The asset identifier for the native asset (PINT).

type AssetId: Parameter + Member + MaybeSerializeDeserialize

Type used to identify the assets.

type Time: Time

Type to keep track of timestamped values

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type WeightInfo: WeightInfo

The weight for this pallet’s extrinsics.